Refactor Navigation components
