+++ src/app/_components/Navigation/DesktopNavigation/DesktopNavigation.tsx
@@ -1,11 +1,8 @@
 'use client'
 
-import Link from 'next/link'
 import { usePathname } from 'next/navigation'
 
-import { ArrowUpRight } from '@phosphor-icons/react'
 import { clsx } from 'clsx'
-import type { Route } from 'next'
 
 import { PATHS } from '@/constants/paths'
 
@@ -13,92 +10,10 @@ import { desktopNavigationItems } from '@/data/navigationData'
 
 import { useNavigationItems } from '@/hooks/useNavigationItems'
 
-import { Icon } from '@/components/Icon'
 import { NavigationPopover } from '@/components/NavigationPopover'
 
-type MainNavItemProps = {
-  href: Route
-  label: string
-  isActive?: boolean
-}
-
-export type SubNavItemProps = {
-  href: string | Route
-  label: string
-  description?: string
-  linkType?: 'internal' | 'externalPrimary' | 'externalSecondary'
-}
-
-function SubNavItem({
-  href,
-  label,
-  description,
-  linkType = 'internal',
-}: SubNavItemProps) {
-  const external = linkType !== 'internal'
-
-  const baseStyles = 'focus:brand-outline group w-full rounded-lg'
-  const extendedStyles = {
-    internal: 'inline-block bg-brand-800 p-4 hover:bg-brand-700',
-    externalPrimary:
-      'inline-block border border-brand-500 bg-brand-700 p-4 hover:border-brand-400 focus:border-transparent',
-    externalSecondary:
-      'inline-flex items-center justify-center gap-1 bg-brand-800 px-3 py-5 text-brand-300 hover:bg-brand-700',
-  }
-
-  const linkClasses = clsx(baseStyles, extendedStyles[linkType])
-  const commonProps = {
-    className: linkClasses,
-    'aria-label': `${label} page (${external ? 'external link' : 'internal link'})`,
-  }
-
-  return external ? (
-    <a href={href} {...commonProps} rel="noopener noreferrer">
-      <div className="inline-flex items-center gap-1">
-        <p className="font-bold">{label}</p>
-        <span className="text-brand-400 group-hover:text-brand-100">
-          <Icon component={ArrowUpRight} size={20} />
-        </span>
-      </div>
-      {description && <p className="mt-4 text-brand-300">{description}</p>}
-    </a>
-  ) : (
-    <Link href={href as Route} {...commonProps}>
-      <p className="mb-1 font-bold">{label}</p>
-      <p className="text-brand-300">{description}</p>
-    </Link>
-  )
-}
-
-const touchTargetMainNavItem = {
-  width: 'px-4',
-  offsetWidth: '-mr-4',
-}
-
-function getMainNavItemStyles(isActive: boolean, isPopover = false) {
-  const baseStyles =
-    'focus:brand-outline rounded-xl py-1.5 text-base hover:bg-brand-700'
-
-  const extendedStyles = isPopover
-    ? 'inline-flex items-center gap-2 pl-4 pr-3 ui-open:bg-brand-700 ui-open:text-brand-400'
-    : clsx('inline-block', touchTargetMainNavItem.width)
-
-  return clsx(
-    baseStyles,
-    extendedStyles,
-    isActive ? 'text-brand-400' : 'text-brand-300',
-  )
-}
-
-function MainNavItem({ label, href, isActive = false }: MainNavItemProps) {
-  return (
-    <li>
-      <Link href={href} className={getMainNavItemStyles(isActive)}>
-        {label}
-      </Link>
-    </li>
-  )
-}
+import { TOUCH_TARGET_MAIN_NAV_ITEM, MainNavItem } from './MainNavItem'
+import { SubNavItem } from './SubNavItem'
 
 const { getInvolvedItems, communityItems, resourcesItems, learnMoreItem } =
   desktopNavigationItems
@@ -123,7 +38,7 @@ export function DesktopNavigation() {
       aria-label="Navigation items"
       className={clsx(
         'relative z-10 hidden lg:flex lg:items-center lg:gap-0.5',
-        touchTargetMainNavItem.offsetWidth,
+        TOUCH_TARGET_MAIN_NAV_ITEM.offsetWidth,
       )}
     >
       <MainNavItem
@@ -142,6 +57,7 @@ export function DesktopNavigation() {
             {getInvolvedInternalItems.map((item) => (
               <SubNavItem key={item.href} {...item} linkType="internal" />
             ))}
+            <SubNavItem {...learnMoreItem} linkType="externalSecondary" />
           </div>
           <div className="space-y-4">
             {getInvolvedExternalItems.map((item) => (
@@ -180,5 +96,20 @@ export function DesktopNavigation() {
         </div>
       </NavigationPopover>
     </ul>
+  )
+}
+
+function getMainNavItemStyles(isActive: boolean, isPopover = false) {
+  const baseStyles =
+    'focus:brand-outline rounded-xl py-1.5 text-base hover:bg-brand-700'
+
+  const extendedStyles = isPopover
+    ? 'inline-flex items-center gap-2 pl-4 pr-3 ui-open:bg-brand-700 ui-open:text-brand-400'
+    : clsx('inline-block', TOUCH_TARGET_MAIN_NAV_ITEM.width)
+
+  return clsx(
+    baseStyles,
+    extendedStyles,
+    isActive ? 'text-brand-400' : 'text-brand-300',
   )
 }
