+++ src/app/_components/Navigation/Navigation.tsx
@@ -2,9 +2,10 @@ import Link from 'next/link'
 
 import { PATHS } from '@/constants/paths'
 
-import { DesktopNavigation } from '@/components/DesktopNavigation'
 import { Logo } from '@/components/Logo'
-import { MobileNavigation } from '@/components/MobileNavigation'
+
+import { DesktopNavigation } from './DesktopNavigation/DesktopNavigation'
+import { MobileNavigation } from './MobileNavigation/MobileNavigation'
 
 export function Navigation() {
   return (
